Many initial updates for my setup
Added a Readme file.

Updated the header information placing a notification about the fork.

Replaced the $nsupdateCommands variable to a set that was working for
me.

Removed the IP and KEY required parameters for the time being.

Added more output information on the duplicate IP checking section -
previous was to exit and I wasn't sure if the script even ran.

Cleaned up the commented sections.